requirements.txt
asgiref==3.5.2
astroid==2.12.4
asttokens==2.1.0
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.10.0
braintree==4.17.1
certifi==2021.10.8
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
comtypes==1.1.11
constantly==15.1.0
cryptography==37.0.4
cycler==0.11.0
daphne==3.0.2
debugpy==1.6.3
decorator==5.1.1
dill==0.3.5.1
distlib==0.3.5
Django==4.1.2
django-autoslug==1.9.8
django-tinymce==3.4.0
djangorestframework==3.14.0
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
filelock==3.7.1
Flask==2.2.2
fonttools==4.37.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipykernel==6.17.1
ipython==8.6.0
isort==5.10.1
itsdangerous==2.1.2
jaraco.context==4.1.1
jedi==0.18.2
Jinja2==3.1.2
jsonify==0.5
jupyter_client==7.4.7
jupyter_core==5.0.0
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
more-itertools==8.12.0
MouseInfo==0.1.3
msgpack==1.0.4
mysql-connector-python==8.0.31
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.23.2
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pipenv==2022.10.25
pipreqs==0.4.11
platformdirs==2.5.2
prompt-toolkit==3.0.33
protobuf==3.20.1
psutil==5.9.4
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.13.0
pyjokes==0.6.0
PyJWT==2.3.0
pylint==2.15.0
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.1
pywin32==303
pyzmq==24.0.1
requests==2.27.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
stack-data==0.6.1
stripe==4.1.0
tomli==2.0.1
tomlkit==0.11.4
tornado==6.2
traitlets==5.5.0
twilio==7.8.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2021.5
urllib3==1.26.9
virtualenv==20.16.3
virtualenv-clone==0.5.7
wcwidth==0.2.5
Werkzeug==2.2.2
wikipedia==1.4.0
winshell==0.6
wolframalpha==5.0.0
wrapt==1.14.1
xmltodict==0.12.0
yarg==0.1.9
zope.interface==5.4.0
